Fix type for shape2str function to use size_t instead of unsigned long

+++ src/tensor/xtensor_lib.cpp
@@ -6,7 +6,7 @@
 #include "tensor/xtensor_lib.h"
 
 
-string shape2str(xt::svector<unsigned long> vec){
+string shape2str(xt::svector<size_t> vec){
     stringstream ss;
     ss << "(";
     for(int idx=0; idx < vec.size(); idx++){
